python/requirements.txt
gymnasium[classic-control,box2d]==0.29.1
torch>=2.2.0
numpy>=1.26.0
Pillow>=10.0.0
pydantic>=2.5.0
